sending authorization token on necessary requests

+++ Front-end/src/services/Auth.js
@@ -8,9 +8,9 @@ export default {
         console.log('credentials recus', credentials)
         return api.post('login', credentials)
     },
-    deleteUser(id) {
+    deleteUser(id, token) {
         const path = 'profile/' + id.id
-        return api.delete(path)
+        return api.delete(path, {headers: {'Authorization': `Bearer ${token}`}})
     }
 }
