+++ Front-end/src/services/Comments.js
@@ -1,13 +1,13 @@
 import { api } from 'boot/axios'
 
 export default {
-    deleteComment(id) {
+    deleteComment(id, token) {
         const path = 'comment/' + id.id
-        return api.delete(path)
+        return api.delete(path, {headers: {'Authorization': `Bearer ${token}`}})
     },
-    createComment(content) {
+    createComment(content, token) {
         console.log('content :', content)
         const path = 'post/' + content.postId
-        return api.post(path , content)
+        return api.post(path , content, {headers: {'Authorization': `Bearer ${token}`}})
     }
 }
