+++ Front-end/src/services/Posts.js
@@ -4,19 +4,19 @@ export default {
     getFeed() {
         return api.get('feed')
     },
-    createPost(content) {
-        return api.post('post', content)
+    createPost(content, token) {
+        return api.post('post', content, {headers: {'Authorization': `Bearer ${token}`}})
     },
     getPost(id) {
         const path = 'post/' + id
         return api.get(path)
     },
-    editPost(id) {
+    editPost(id, token) {
         const path = 'post/edit/' + id
-        return api.put(path)
+        return api.put(path, {headers: {'Authorization': `Bearer ${token}`}})
     },
-    deletePost(id) {
+    deletePost(id, token) {
         const path = 'post/' + id
-        return api.delete(path)
+        return api.delete(path, {headers: {'Authorization': `Bearer ${token}`}})
     }
 }
